[glsl] Promote glsl-new to glsl
