+++ Cargo.toml
@@ -15,7 +15,6 @@ fxhash = "0.2"
 log = "0.4"
 num-traits = "0.2"
 spirv = { package = "spirv_headers", version = "1.4.2", optional = true }
-glsl = { version = "4.1", optional = true }
 pomelo = { version = "0.1.4", optional = true }
 thiserror = "1.0"
 serde = { version = "1.0", features = ["derive"], optional = true }
@@ -23,8 +22,7 @@ petgraph = { version ="0.5", optional = true }
 
 [features]
 default = []
-glsl_preprocessor = ["glsl"]
-glsl-new = ["pomelo"]
+glsl = ["pomelo"]
 glsl-validate = []
 glsl-out = []
 serialize = ["serde"]
