+++ examples/convert.rs
@@ -29,9 +29,6 @@ fn main() {
         println!("Call with <input> <output>");
         return;
     }
-    #[cfg(any(feature = "glsl", feature = "glsl-new"))]
-    let prefer_glsl_new =
-        !cfg!(feature = "glsl") || env::var("PREFER_GLSL_NEW").unwrap_or_default() == "1";
     let module = match Path::new(&args[1])
         .extension()
         .expect("Input has no extension?")
@@ -48,73 +45,23 @@ fn main() {
             let input = fs::read_to_string(&args[1]).unwrap();
             naga::front::wgsl::parse_str(&input).unwrap()
         }
-        #[cfg(any(feature = "glsl", feature = "glsl-new"))]
+        #[cfg(feature = "glsl")]
         "vert" => {
             let input = fs::read_to_string(&args[1]).unwrap();
-            let mut module: Option<naga::Module> = None;
-            if prefer_glsl_new {
-                #[cfg(feature = "glsl-new")]
-                {
-                    module = Some(
-                        naga::front::glsl_new::parse_str(
-                            &input,
-                            "main".to_string(),
-                            naga::ShaderStage::Vertex,
-                        )
-                        .unwrap(),
-                    )
-                }
-            }
-            if module.is_none() {
-                #[cfg(feature = "glsl")]
-                {
-                    module = Some(
-                        naga::front::glsl::parse_str(
-                            &input,
-                            "main".to_string(),
-                            naga::ShaderStage::Vertex,
-                        )
-                        .unwrap(),
-                    )
-                }
-            }
-            module.unwrap()
+            naga::front::glsl::parse_str(&input, "main".to_string(), naga::ShaderStage::Vertex)
+                .unwrap()
         }
-        #[cfg(any(feature = "glsl", feature = "glsl-new"))]
+        #[cfg(feature = "glsl")]
         "frag" => {
             let input = fs::read_to_string(&args[1]).unwrap();
-            let mut module = None;
-            if prefer_glsl_new {
-                #[cfg(feature = "glsl-new")]
-                {
-                    module = Some(
-                        naga::front::glsl_new::parse_str(
-                            &input,
-                            "main".to_string(),
-                            naga::ShaderStage::Fragment {
-                                early_depth_test: None,
-                            },
-                        )
-                        .unwrap(),
-                    )
-                }
-            }
-            if module.is_none() {
-                #[cfg(feature = "glsl")]
-                {
-                    module = Some(
-                        naga::front::glsl::parse_str(
-                            &input,
-                            "main".to_string(),
-                            naga::ShaderStage::Fragment {
-                                early_depth_test: None,
-                            },
-                        )
-                        .unwrap(),
-                    )
-                }
-            }
-            module.unwrap()
+            naga::front::glsl::parse_str(
+                &input,
+                "main".to_string(),
+                naga::ShaderStage::Fragment {
+                    early_depth_test: None,
+                },
+            )
+            .unwrap()
         }
         #[cfg(feature = "glsl")]
         "comp" => {
