+++ src/front/glsl/rosetta_tests.rs
@@ -9,7 +9,7 @@ fn rosetta_test(file_name: &str, stage: crate::ShaderStage) {
     let expected =
         fs::read_to_string(test_dir.join(file_name).with_extension("expected.ron")).unwrap();
 
-    let module = crate::front::glsl_new::parse_str(&input, "main".to_string(), stage).unwrap();
+    let module = super::parse_str(&input, "main".to_string(), stage).unwrap();
     let output = ron::ser::to_string_pretty(&module, Default::default()).unwrap();
 
     assert_diff!(output.as_str(), expected.as_str(), "", 0);
